# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Unicis Platform API
  description: |
    REST API for the Unicis GRC (Governance, Risk & Compliance) platform.

    ## Authentication
    All endpoints require authentication via one of:
    - **Bearer token** — API key from team settings (`Authorization: Bearer <key>`)
    - **Session cookie** — browser session via NextAuth

    ## Response format
    Every response uses a consistent envelope:
    ```json
    { "data": <payload>, "error": null }       // success
    { "error": { "message": "..." } }          // error
    ```

    ## Roles & Permissions
    API keys authenticate with **ADMIN** role permissions.
    Available roles: `OWNER`, `ADMIN`, `MEMBER`, `AUDITOR`.
  version: 0.2.2
  contact:
    name: Unicis Tech
    url: https://unicis.tech
servers:
  - url: '{baseUrl}'
    variables:
      baseUrl:
        default: 'http://localhost:4002'
        description: Base URL of the Unicis Platform instance
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from team settings
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (browser-based)
  parameters:
    slug:
      name: slug
      in: path
      required: true
      description: Team slug identifier
      schema:
        type: string
        example: my-team
    taskNumber:
      name: taskNumber
      in: path
      required: true
      description: Task number within the team
      schema:
        type: integer
        example: 1
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
          required:
            - message
    EmptyDataResponse:
      type: object
      properties:
        data:
          type: object
        error:
          nullable: true
    Task:
      type: object
      properties:
        id:
          type: integer
        taskNumber:
          type: integer
        teamId:
          type: string
        authorId:
          type: string
        title:
          type: string
        status:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        duedate:
          type: string
        description:
          type: string
        properties:
          type: object
          description: Flexible JSON properties (controls, risk, assessments, etc.)
        comments:
          type: array
          deprecated: true
          description: Legacy embedded comments. Use the dedicated task comments endpoint instead.
          items:
            $ref: '#/components/schemas/Comment'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
    TaskPriority:
      type: string
      enum:
        - low
        - medium
        - high
    Comment:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        taskId:
          type: integer
        createdById:
          type: string
        createdBy:
          type: object
          properties:
            name:
              type: string
              nullable: true
            email:
              type: string
...
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: integer
        filename:
          type: string
        url:
          type: string
...
    RiskInput:
      type: object
      properties:
        prevRisk:
          type: array
          items:
            type: object
          description: Previous risk assessment data
        nextRisk:
          type: array
          items:
            type: object
          description: New risk assessment data
      required:
        - prevRisk
        - nextRisk
    ProcedureInput:
      type: object
      properties:
        prevProcedure:
          type: array
          items:
            type: object
          description: Previous procedure data
        nextProcedure:
          type: array
          items:
            type: object
          description: New procedure data
      required:
        - prevProcedure
        - nextProcedure
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: Unauthorized
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: You are not allowed to perform this action
...
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/teams/{slug}/tasks/{taskNumber}/rm:
    parameters:
      - $ref: '#/components/parameters/slug'
      - $ref: '#/components/parameters/taskNumber'

    post:
      operationId: saveRiskManagement
      tags: [Risk Management (RM)]
      summary: Save risk assessment for a task
      description: Creates or updates the risk management assessment. Requires `task:update` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskInput'
      responses:
        '200':
          description: Risk saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      task:
                        $ref: '#/components/schemas/Task'
                  error:
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      operationId: deleteRiskManagement
      tags: [Risk Management (RM)]
      summary: Delete risk assessment from a task
      description: Removes the risk management assessment. Requires `task:update` permission.
      responses:
        '200':
          description: Risk deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── Privacy Impact Assessment (PIA) ─────────────────────
  /api/teams/{slug}/tasks/{taskNumber}/pia:
    parameters:
      - $ref: '#/components/parameters/slug'
      - $ref: '#/components/parameters/taskNumber'

    post:
      operationId: savePrivacyImpact
      tags: [Privacy Impact (PIA)]
      summary: Save privacy impact assessment for a task
      description: Creates or updates the privacy impact assessment. Requires `task:update` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskInput'
      responses:
        '200':
          description: PIA saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      task:
                        $ref: '#/components/schemas/Task'
                  error:
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      operationId: deletePrivacyImpact
      tags: [Privacy Impact (PIA)]
      summary: Delete privacy impact assessment from a task
      description: Removes the privacy impact assessment. Requires `task:update` permission.
      responses:
        '200':
          description: PIA deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── Record of Processing Activities (RPA) ───────────────
  /api/teams/{slug}/tasks/{taskNumber}/rpa:
    parameters:
      - $ref: '#/components/parameters/slug'
      - $ref: '#/components/parameters/taskNumber'

    post:
      operationId: saveRecordOfProcessing
      tags: [Record of Processing (RPA)]
      summary: Save record of processing for a task
      description: Creates or updates the record of processing activities. Requires `task:update` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcedureInput'
      responses:
        '200':
          description: RPA saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      operationId: deleteRecordOfProcessing
      tags: [Record of Processing (RPA)]
      summary: Delete record of processing from a task
      description: Removes the record of processing activities. Requires `task:update` permission.
      responses:
        '200':
          description: RPA deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── Transfer Impact Assessment (TIA) ────────────────────
  /api/teams/{slug}/tasks/{taskNumber}/tia:
    parameters:
      - $ref: '#/components/parameters/slug'
      - $ref: '#/components/parameters/taskNumber'

    post:
      operationId: saveTransferImpact
      tags: [Transfer Impact (TIA)]
      summary: Save transfer impact assessment for a task
      description: Creates or updates the transfer impact assessment. Requires `task:update` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcedureInput'
      responses:
        '200':
          description: TIA saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      operationId: deleteTransferImpact
      tags: [Transfer Impact (TIA)]
      summary: Delete transfer impact assessment from a task
      description: Removes the transfer impact assessment. Requires `task:update` permission.
      responses:
        '200':
          description: TIA deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── API Keys ─────────────────────────────────────────────
...