# pages/api/teams/[slug]/tasks/[taskNumber]/pia.ts
import type { NextApiRequest, NextApiResponse } from 'next';
import { throwIfNoTeamAccess } from 'models/team';
import { throwIfNotAllowed } from 'models/user';
import { deleteRisk, saveRisk } from 'models/pia';
import { trackServerEvent } from '@/lib/matomo/server';
import { MatomoEvent } from '@/lib/matomo/events';
import { validateApiRequestBody } from '@/lib/api-validation';
import { piaWriteRequestSchema } from '@/lib/properties';

export default async function handler(
  req: NextApiRequest,
  res: NextApiResponse
) {
  const { method } = req;

  switch (method) {
    case 'POST':
      return handlePOST(req, res);
    case 'DELETE':
      return handleDELETE(req, res);
    default:
      res.setHeader('Allow', ['POST', 'DELETE']);
      res.status(405).json({
        data: null,
        error: { message: `Method ${method} Not Allowed` },
      });
  }
}

const handlePOST = async (req: NextApiRequest, res: NextApiResponse) => {
  const teamMember = await throwIfNoTeamAccess(req, res);
  throwIfNotAllowed(teamMember, 'task', 'update');

  const { slug, taskNumber } = req.query;

  const taskNumberAsNumber = Number(taskNumber);

  if (isNaN(taskNumberAsNumber)) {
    return res.status(400).json({
      error: {
        message: 'Invalid task number',
      },
    });
  }

  const body = validateApiRequestBody(piaWriteRequestSchema, req.body, res);

  if (!body) {
    return;
  }

  const { nextRisk } = body;

  const task = await saveRisk({
    user: teamMember.user,
    taskNumber: taskNumberAsNumber,
    slug: slug as string,
    nextRisk,
  });

  if (!task) {
    return res.status(400).json({
      error: {
        message: 'Something went wrong!',
      },
    });
  }

  trackServerEvent(MatomoEvent.PiaSaved);

  return res.status(200).json({ data: { task }, error: null });
};

const handleDELETE = async (req: NextApiRequest, res: NextApiResponse) => {
  const teamMember = await throwIfNoTeamAccess(req, res);
  throwIfNotAllowed(teamMember, 'task', 'update');

  const { slug, taskNumber } = req.query;

  const taskNumberAsNumber = Number(taskNumber);

  if (isNaN(taskNumberAsNumber)) {
    return res.status(400).json({
      error: {
        message: 'Invalid task number',
      },
    });
  }

  const task = await deleteRisk({
    user: teamMember.user,
    taskNumber: taskNumberAsNumber,
    slug: slug as string,
  });

  if (!task) {
    return res.status(400).json({
      error: {
        message: 'Something went wrong!',
      },
    });
  }

  return res.status(200).json({ data: {}, error: null });
};